Improvment: Updating Readme and cosmetic changes.
This commit contains changes which,
- Update the Readme with detailed steps to train and run inference.
- Cosmetic changes to sanitize the repository.
